fix: Criando arquivo de utils para reutilização posterior em vários arquivos
